Revert "Merge branch 'parser-tweaks' into dev"
This reverts commit 109dbbcaa6c3f7df12d99d836132458ace03d356, reversing
changes made to 7c4f33e1c4371c8fc605b32568917a892b8edd06.
