+++ snorkel/models/candidate.py
@@ -80,10 +80,6 @@ class Candidate(SnorkelBase):
         """Get a tuple of the consituent contexts making up this candidate"""
         return tuple(getattr(self, name) for name in self.__argnames__)
 
-    def get_parent(self):
-        # Assumes all arguments have the same parent
-        return self.get_contexts()[0].parent
-
     def get_cids(self):
         """Get a tuple of the canonical IDs (CIDs) of the contexts making up this candidate"""
         return tuple(getattr(self, name + "_cid") for name in self.__argnames__)
